VectoreStoreProject1/requirements.txt
langchain==0.2.15
langchain-core==0.2.36
langchain-community==0.2.14
langchain-openai==0.1.23
langchain-chroma==0.1.3
chromadb==0.5.7
openai==1.63.0
python-dotenv==1.0.1
